feat: implement phase 1 workflow improvements
Make structured oracle the default comparison tool and add
comprehensive workflow documentation to reduce manual debugging time.

Changes:
* Rename oracle.js to oracle-simple.js (legacy string comparison)
* Rename oracle-structured.js to oracle.js (new default)
* Add workflow-test.sh wrapper for structured oracle + batch analysis
* Create docs/RENDERING_WORKFLOW.md with step-by-step guide covering:
  - When to use each oracle script
  - How to interpret structured diff output
  - Common failure patterns and fixes
  - Advanced debugging techniques
* Update CLAUDE.md test commands to reference new defaults

Expected impact: 50% reduction in manual debugging time per style.

Closes #21, #22, #23
